ai-legal-assistant/requirements.txt
streamlit
PyPDF2
pdfplumber
reportlab
groq
google-generativeai
python-dotenv
requests
langchain
langchain-community
langchain-google-genai
langchain-huggingface
sentence-transformers
faiss-cpu
